update question serializing and sync question_data id with question_meta id

+++ modules/core/src/main/java/org/vstu/compprehension/models/repository/QuestionDataRepository.java
@@ -3,5 +3,5 @@ package org.vstu.compprehension.models.repository;
 import org.springframework.data.repository.CrudRepository;
 import org.vstu.compprehension.models.entities.QuestionDataEntity;
 
-public interface QuestionDataRepository extends CrudRepository<QuestionDataEntity, Long> {
+public interface QuestionDataRepository extends CrudRepository<QuestionDataEntity, Integer> {
 }
